+++ doq/Changelog.md
@@ -2,6 +2,6 @@
 
 <time datetime="2025-08-12">2025-08-12</time>
 
-- Adding functionalit to subnamespace ```types```(StringClass, ListClass)
+- Adding functionality to subnamespace ```types```(StringClass, ListClass)
 - Setting up PHPUnit functionality
 - Adding Unit test with usage of attribute ```TestWith``` as <var>DataProvider</var>